alpaca-mini-broker/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
httpx==0.27.2
python-dotenv==1.0.1
sqlalchemy==2.0.36
aiosqlite==0.20.0
jinja2==3.1.4